Standardize on date_make_date() instead of calling date_create() in some places so there is a central place to test the timezone and date values before creating the date.
Add method to customize the labeling of date parts to allow the label above the date part, inserted as the first item in the options array, or omitted.

Remove some functions that are not being used or move them so they are called only when needed to clean up the code.